Fix unboxing warning
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1547846 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/catalina/tribes/tipis/AbstractReplicatedMap.java
@@ -277,7 +277,7 @@ public abstract class AbstractReplicatedMap<K,V>
             Member[] members = mapMembers.keySet().toArray(new Member[mapMembers.size()]);
             long now = System.currentTimeMillis();
             for (Member member : members) {
-                long access = mapMembers.get(member);
+                long access = mapMembers.get(member).longValue();
                 if ( (now - access) > timeout ) {
                     memberDisappeared(member);
                 }
